# backend/swagger.json
{"openapi":"3.0.0","info":{"title":"Banking API","version":"1.0.0","description":"A simple banking API with authentication, accounts, cards, and transactions"},"servers":[{"url":"http://localhost:3001","description":"Development server"}],"components":{"securitySchemes":{"bearerAuth":{"type":"http","scheme":"bearer","bearerFormat":"JWT"}}},"security":[{"bearerAuth":[]}],"paths":{"/login":{"post":{"summary":"Authenticate user and get JWT tokens","tags":["Authentication"],"requestBody":{"required":true,"content":{"application/json":{"schema":{"type":"object","required":["username","password"],"properties":{"username":{"type":"string","example":"test@test.test"},"password":{"type":"string","example":"password@123"}}}}}},"responses":{"200":{"description":"Login successful","content":{"application/json":{"schema":{"type":"object","properties":{"token":{"type":"string","description":"JWT access token"},"refreshToken":{"type":"string","description":"JWT refresh token"}}}}}},"401":{"description":"Invalid credentials","content":{"application/json":{"schema":{"type":"object","properties":{"message":{"type":"string","example":"Invalid credentials"}}}}}}}}},"/refresh-token":{"post":{"summary":"Refresh JWT access token using refresh token","tags":["Authentication"],"requestBody":{"required":true,"content":{"application/json":{"schema":{"type":"object","required":["refreshToken"],"properties":{"refreshToken":{"type":"string","description":"JWT refresh token"}}}}}},"responses":{"200":{"description":"Token refreshed successfully","content":{"application/json":{"schema":{"type":"object","properties":{"token":{"type":"string","description":"New JWT access token"},"refreshToken":{"type":"string","description":"New JWT refresh token"}}}}}},"401":{"description":"No refresh token provided"},"403":{"description":"Invalid or expired refresh token"}}}},"/accounts":{"get":{"summary":"Get user accounts with balances","tags":["Accounts"],"security":[{"bearerAuth":[]}],"responses":{"200":{"description":"List of user accounts with balances","content":{"application/json":{"schema":{"type":"array","items":{"type":"object","properties":{"id":{"type":"integer","example":1},"user_id":{"type":"integer","example":1},"name":{"type":"string","example":"Checking"},"balance":{"type":"number","format":"float","example":1950.25}}}}}}},"401":{"description":"Unauthorized - No valid token"},"403":{"description":"Forbidden - Invalid token"},"500":{"description":"Internal server error"}}}},"/cards":{"get":{"summary":"Get user's credit/debit cards","tags":["Cards"],"security":[{"bearerAuth":[]}],"responses":{"200":{"description":"List of user's cards","content":{"application/json":{"schema":{"type":"array","items":{"type":"object","properties":{"id":{"type":"integer","example":1},"user_id":{"type":"integer","example":1},"number":{"type":"string","example":"4111111111111111"},"expiry":{"type":"string","example":"12/26"},"cvv":{"type":"string","example":"123"}}}}}}},"401":{"description":"Unauthorized - No valid token"},"403":{"description":"Forbidden - Invalid token"}}}},"/transactions":{"get":{"summary":"Get user transactions with search, sort, and pagination","tags":["Transactions"],"security":[{"bearerAuth":[]}],"parameters":[{"in":"query","name":"search","schema":{"type":"string"},"description":"Search term for description or type","example":"grocery"},{"in":"query","name":"sort","schema":{"type":"string","enum":["date","amount","type","description"]},"description":"Field to sort by","example":"date"},{"in":"query","name":"order","schema":{"type":"string","enum":["asc","desc"]},"description":"Sort order","example":"desc"},{"in":"query","name":"page","schema":{"type":"integer","minimum":1},"description":"Page number","example":1},{"in":"query","name":"limit","schema":{"type":"integer","minimum":1,"maximum":100},"description":"Number of items per page","example":10}],"responses":{"200":{"description":"List of transactions","content":{"application/json":{"schema":{"type":"array","items":{"type":"object","properties":{"id":{"type":"integer","example":1},"user_id":{"type":"integer","example":1},"account_id":{"type":"integer","example":1},"amount":{"type":"number","format":"float","example":-50.25},"type":{"type":"string","example":"debit"},"description":{"type":"string","example":"Grocery Store"},"date":{"type":"string","format":"date-time","example":"2024-06-01T10:00:00Z"}}}}}}},"401":{"description":"Unauthorized - No valid token"},"403":{"description":"Forbidden - Invalid token"}}}}},"tags":[]}